refactor: restructure project into clean architecture
