+++ internal/ports/order_repository.go
@@ -1,0 +1,7 @@
+package ports
+
+import "ecommerce_order/internal/domain"
+
+type OrderRepository interface {
+	Save(order domain.Order) error
+}
