+++ internal/ports/queue.go
@@ -1,0 +1,7 @@
+package ports
+
+import "ecommerce_order/internal/domain"
+
+type OrderPublisher interface {
+	Publish(order domain.Order) error
+}
